Updates default Win 10 evaluation media to build 1909

+++ Lability.psm1
@@ -33,21 +33,23 @@ Import-LocalizedData @importLocalizedDataParams;
 $moduleRoot = Split-Path -Path $MyInvocation.MyCommand.Path -Parent;
 $moduleSrcPath = Join-Path -Path $moduleRoot -ChildPath 'Src';
 Get-ChildItem -Path $moduleSrcPath -Include *.ps1 -Exclude '*.Tests.ps1' -Recurse |
-ForEach-Object {
-    Write-Verbose -Message ('Importing library\source file ''{0}''.' -f $_.FullName);
-    ## https://becomelotr.wordpress.com/2017/02/13/expensive-dot-sourcing/
-    . ([System.Management.Automation.ScriptBlock]::Create(
-            [System.IO.File]::ReadAllText($_.FullName)
-        ));
-}
+    ForEach-Object {
+        Write-Verbose -Message ('Importing library\source file ''{0}''.' -f $_.FullName);
+        # https://becomelotr.wordpress.com/2017/02/13/expensive-dot-sourcing/
+        . ([System.Management.Automation.ScriptBlock]::Create(
+                [System.IO.File]::ReadAllText($_.FullName)
+            ));
+    }
 
 ## Deploy builtin certificates to %ALLUSERSPROFILE%\PSLab
 $moduleConfigPath = Join-Path -Path $moduleRoot -ChildPath 'Config';
 $allUsersConfigPath = Join-Path -Path $env:AllUsersProfile -ChildPath "$($labDefaults.ModuleName)\Certificates\";
 [ref] $null = New-Directory -Path $allUsersConfigPath;
 Get-ChildItem -Path $moduleConfigPath -Include *.cer, *.pfx -Recurse | ForEach-Object {
-    Write-Verbose -Message ('Updating certificate ''{0}''.' -f $_.FullName);
-    Copy-Item -Path $_ -Destination $allUsersConfigPath;
+    if (-not (Test-Path -Path (Join-Path -Path $allUsersConfigPath -ChildPath $_.Name))) {
+        Write-Verbose -Message ('Updating certificate ''{0}''.' -f $_.FullName);
+        Copy-Item -Path $_ -Destination $allUsersConfigPath;
+    }
 }
 
 ## Create the credential check scriptblock
